crm-backend/requirements-dev.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
requests==2.31.0
openpyxl==3.1.5
reportlab==4.0.7
python-dateutil==2.8.2
APScheduler==3.10.4
alembic==1.13.0
